UPDATE CHANGE PASSWORD + PROFILE
CMS Can change password and profile

+++ admin/users/index.php
@@ -60,25 +60,35 @@ $query = $sql->select();
 										?>
 									</td>
 									<td class="text-center">
-										<a href="<?=URL?>admin/users/password.php?page=<?=$_GET["page"]?>&id=<?=$result["id"]?>" class="btn btn-info btn-sm text-white" title="เปลี่ยนรหัสผ่าน">
-											<i class="fa fa-key"></i>
-										</a>
+										<?php
+										if( empty($result["is_owner"]) ){
+											?>
+											<a href="<?=URL?>admin/users/password.php?page=<?=$_GET["page"]?>&id=<?=$result["id"]?>" class="btn btn-info btn-sm text-white" title="เปลี่ยนรหัสผ่าน">
+												<i class="fa fa-key"></i>
+											</a>
+											<?php 
+										} 
+										?>
 
 										<a href="<?=URL?>admin/users/forms.php?page=<?=$_GET["page"]?>&id=<?=$result["id"]?>" class="btn btn-warning btn-sm text-white" title="แก้ไขข้อมูล">
 											<i class="fa fa-pen"></i>
 										</a>
 
 										<?php 
-										$ops = [
-											"title" => "ยืนยันการลบข้อมูล",
-											"text" => "คุณต้องการลบข้อมูล ".$result["name"]. "หรือไม่ ?",
-											"btnconfirm" => "btn btn-danger m-1",
-											"textconfirm" => "ลบข้อมูล"
-										];
+										if( empty($result["is_owner"]) ){
+											$ops = [
+												"title" => "ยืนยันการลบข้อมูล",
+												"text" => "คุณต้องการลบข้อมูล ".$result["name"]. "หรือไม่ ?",
+												"btnconfirm" => "btn btn-danger m-1",
+												"textconfirm" => "ลบข้อมูล"
+											];
+											?>
+											<a href="<?=URL?>admin/users/delete.php?page=<?=$_GET["page"]?>&id=<?=$result["id"]?>" class="btn btn-danger btn-confirm btn-sm" data-title="ยืนยันการลบข้อมูล" data-options="<?=stringify($ops)?>">
+												<i class="fa fa-trash"></i>
+											</a>
+											<?php 
+										} 
 										?>
-										<a href="<?=URL?>admin/users/delete.php?page=<?=$_GET["page"]?>&id=<?=$result["id"]?>" class="btn btn-danger btn-confirm btn-sm" data-title="ยืนยันการลบข้อมูล" data-options="<?=stringify($ops)?>">
-											<i class="fa fa-trash"></i>
-										</a>
 									</td>
 								</tr>
 								<?php
